// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  customer: 'customer',
  shop: 'shop',
  admin: 'admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipping: 'shipping',
  cancelled: 'cancelled'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const Rating = {
  one: 'one',
  two: 'two',
  three: 'three',
  four: 'four',
  five: 'five'
} as const

export type Rating = (typeof Rating)[keyof typeof Rating]